requirements.txt
faiss-cpu
sentence-transformers
pypdf
tqdm
numpy
pandas
